封装MySql操作,创建对应方法.

+++ admin/message/message_delete.php
@@ -5,13 +5,19 @@
  * Date: 2017/12/1
  * Time: 9:57
  */
-include('../../books/common/db.php');
-$id=$_GET['id'];
-$sql="DELETE FROM book_message WHERE id='$id'";
-$result=$mysqli->query($sql);
-if(isset($result)){
-    $mysqli->close();
-    header("location:message.php");
+header("Content-Type: text/html; charset=utf-8");
+session_start();
+if(isset($_SESSION['user'])) {
+    include('../../books/common/db.php');
+    $id = $_GET['id'];
+    $sql = "DELETE FROM book_message WHERE id='$id'";
+    $result = $mysqli->query($sql);
+    if (isset($result)) {
+        $mysqli->close();
+        header("location:message.php");
+    } else {
+        echo "no error!!";
+    }
 }else{
-    echo "no error!!";
+    echo 'sorry,你还没有登录';
 }
